started moving w_near specific functionality into its own rs file
